Adding get_fertility in tile_type

+++ model/crop_type.py
@@ -30,8 +30,8 @@ class CropType(Enum):
     def get_seed_price(self) -> float:
         return float(self.config[f"croptype.{self.name.lower()}.seedprice"])
 
-    def get_growth_time(self) -> float:
-        return float(self.config[f"croptype.{self.name.lower()}.growthtime"])
+    def get_growth_time(self) -> int:
+        return int(self.config[f"croptype.{self.name.lower()}.growthtime"])
 
     def get_fertility_sensitivity(self) -> float:
         return float(self.config[f"croptype.{self.name.lower()}.fertilitysens"])
